Refactors link structure, most pages accessible
